% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{db_append}
\alias{db_append}
\title{SQL Database Append}
\usage{
db_append(path, tbl, data)
}
\arguments{
\item{path}{path to SQL database}

\item{tbl}{specify a tibble name within the SQL database}

\item{data}{the dataframe to be saved}
}
\description{
This is a helper function that will write a dataframe to a SQL database
}
