% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{db_get_data}
\alias{db_get_data}
\title{SQL Database Retrieve}
\usage{
db_get_data(tbl_dat, path = "sql_data/omdata.db")
}
\arguments{
\item{tbl_dat}{which table from the SQL database do you want to retrieve}

\item{path}{path to database}
}
\description{
This is a helper function that will retreive a dataframe to a SQL database
}
