% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{db_remove}
\alias{db_remove}
\title{SQL Database Remove}
\usage{
db_remove(path, datasets = NULL, remove_cleaned_data = T)
}
\arguments{
\item{path}{path to database}

\item{datasets}{which table from the SQL database do you want to remove}

\item{remove_cleaned_data}{boolean remove all datasets that are created through the cleaning script}
}
\description{
This is a helper function that will remove a dataframe from a SQL database
}
