% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{msg}
\alias{msg}
\title{Send a fancy message}
\usage{
msg(type, type_style = crayon::make_style("red4"), msg)
}
\arguments{
\item{type}{what message should be displayed in the beginning}

\item{type_style}{crayon color or style}

\item{msg}{what message should be printed}
}
\description{
Print a beautiful message in the terminal
}
