% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{perspective_api_key}
\alias{perspective_api_key}
\title{Check if API key is present}
\usage{
perspective_api_key(test = F)
}
\arguments{
\item{test}{necessary when in a test environment. Defaults to \code{FALSE}.}
}
\description{
Check if API key is present
}
