% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prsp.R
\name{print_progress}
\alias{print_progress}
\title{Print progress in purrr::imap environment}
\usage{
print_progress(x, total)
}
\arguments{
\item{x}{interator number.}

\item{total}{length of items to be iterated over.}
}
\value{
a \code{chr}
}
\description{
Provide iterator number and total length of items to be iterated over
}
