% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prsp.R
\name{prsp_stream}
\alias{prsp_stream}
\title{Analyze comments with Perspective API}
\usage{
prsp_stream(.data, text = NULL, text_id = NULL, ..., safe_output = F,
  verbose = F)
}
\arguments{
\item{.data}{a dataset with a text and text_id column.}

\item{text}{a character vector with text you want to score.}

\item{text_id}{a unique ID for the text that you supply}

\item{...}{arguments passed to \code{\link{prsp_score}}.}

\item{safe_output}{wraps the function into a \code{purrr::safely} environment (defaults to \code{FALSE}). Loop will run without pause and catch + output errors in a tidy \code{tibble} along with the results.}

\item{verbose}{narrates the streaming procedure (defaults to \code{FALSE}).}
}
\value{
a \code{tibble}
}
\description{
Provide a character string with your text, your API key and what scores you want to obtain.
}
\details{
For more details see ?peRspective or \href{https://github.com/conversationai/perspectiveapi/blob/master/api_reference.md}{Perspective API documentation}
}
