% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{specify_decimal}
\alias{specify_decimal}
\title{Specify a decimal}
\usage{
specify_decimal(x, k)
}
\arguments{
\item{x}{a number to be rounded}

\item{k}{round to which position after the comma}
}
\description{
Specify a decimal
}
\examples{
## specify 2 decimals of a number
specify_decimal(1.0434, 2)
}
