% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prsp_score.R
\name{unnest_scores}
\alias{unnest_scores}
\title{Unnest scores coming out of Perspective API}
\usage{
unnest_scores(Output, score_model, score_sentences, text)
}
\arguments{
\item{Output}{comes out of the \code{GET} call.}

\item{score_model}{Specify what model do you want to use (for example \code{TOXICITY} and/or \code{SEVERE_TOXICITY}). Specify a character vector if you want more than one score. See \code{peRspective::prsp_models}.}

\item{score_sentences}{A boolean value that indicates if the request should return spans that describe the scores for each part of the text (currently done at per sentence level). Defaults to \code{FALSE}.}

\item{text}{a character string.}
}
\value{
a \code{tibble}
}
\description{
For more details see \code{?peRspective} or \href{https://developers.perspectiveapi.com/s/docs}{Perspective API documentation}
}
