% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{db_append}
\alias{db_append}
\title{SQL Database Append}
\usage{
db_append(path, tbl, data)
}
\arguments{
\item{path}{path to SQL database}

\item{tbl}{name of the table in SQL database}

\item{data}{the object dataframe that goes into the SQL database}
}
\description{
This is a helper function that will write a dataframe to a SQL database
}
