% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prsp.R
\docType{package}
\name{perspective-package}
\alias{perspective-package}
\alias{peRspective}
\title{peRspective: Interface to the Perspective API}
\description{
Provides access to the Perspective API (\url{http://www.perspectiveapi.com/}).
}
\section{Setup}{

\enumerate{
\item Create a Google Cloud project in your \href{https://console.developers.google.com/}{Google Cloud console}
\item Go to \href{https://console.developers.google.com/apis/api/commentanalyzer.googleapis.com/overview}{Perspective API's overview page} and click \strong{\emph{Enable}}
\item Go to the \href{https://console.developers.google.com/apis/credentials}{API credentials page}, just click \strong{\emph{Create credentials}}, and choose "API Key".
}
}

\section{API}{

\pkg{peRspective} functions will read the API key from
environment variable \code{perspective_api_key}.
You can specify it like this at the start of your script:

\code{Sys.setenv(perspective_api_key = "**********")}

To start R session with the
initialized environvent variable create an \code{.Renviron} file in your R home
with a line like this:

\code{perspective_api_key = "**********"}

To check where your R home is, try \code{normalizePath("~")}.
}

\section{Quota and character length Limits}{

You can check your quota limits by going to \href{https://console.cloud.google.com/apis/api/commentanalyzer.googleapis.com/quotas}{your google cloud project's Perspective API page}, and check
your projects quota usage at
\href{https://console.cloud.google.com/iam-admin/quotas}{the cloud console quota usage page}.

The maximum text size per request is 3000 bytes.
}

