% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_progress}
\alias{print_progress}
\title{Print progress in purrr::imap environment}
\usage{
print_progress(x, total, print_prct = F)
}
\arguments{
\item{x}{iterator number.}

\item{total}{length of items to be iterated over.}

\item{print_prct}{only print percentage progress (defaults to \code{FALSE}).}
}
\value{
a \code{chr}
}
\description{
Provide iterator number and total length of items to be iterated over
}
\examples{
## Print progress (1 out of 100)
print_progress(1, 100)

## Only print percentage
print_progress(1, 100, print_prct = TRUE)
}
