% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prsp.R
\name{prsp_score}
\alias{prsp_score}
\title{Analyze comments with Perspective API}
\usage{
prsp_score(text, key, score_model)
}
\arguments{
\item{text}{a character string.}

\item{key}{Your API key (see here: \code{https://github.com/conversationai/perspectiveapi/blob/master/quickstart.md}).}

\item{score_model}{Specify what model do you want to use (e.g. \code{TOXICITY} or \code{SEVERE_TOXICITY}). Specify a character vector if you want more than one score. See \code{peRspective::prsp_models}.}
}
\value{
a \code{tibble}
}
\description{
Provide a character string with your text, your API key and what scores you want to obtain.
}
