% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prsp.R
\name{prsp_score}
\alias{prsp_score}
\title{Analyze comments with Perspective API}
\usage{
prsp_score(text, text_id = NULL, languages = NULL,
  score_sentences = F, score_model, sleep = 1, key = NULL)
}
\arguments{
\item{text}{a character string.}

\item{text_id}{a unique ID for the text that you supply (required).}

\item{languages}{A vector of \href{https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes}{ISO 631-1} two-letter language codes specifying the language(s) that comment is in (for example, "en", "es", "fr", "de", etc). If unspecified, the API will autodetect the comment language. If language detection fails, the API returns an error.}

\item{score_sentences}{A boolean value that indicates if the request should return spans that describe the scores for each part of the text (currently done at per sentence level). Defaults to \code{FALSE}.}

\item{score_model}{Specify what model do you want to use (for example \code{TOXICITY} and/or \code{SEVERE_TOXICITY}). Specify a character vector if you want more than one score. See \code{peRspective::prsp_models}.}

\item{sleep}{how long should \code{prsp_score} wait between each call}

\item{key}{Your API key (\href{https://github.com/conversationai/perspectiveapi/blob/master/quickstart.md}{see here} to set up an API key).}
}
\value{
a \code{tibble}
}
\description{
Provide a character string with your text, your API key and what scores you want to obtain.
}
\details{
For more details see \code{?peRspective} or \href{https://github.com/conversationai/perspectiveapi/blob/master/api_reference.md}{Perspective API documentation}
}
